Restrcuture scaling

+++ src/metamodels/PerfectMetamodel.py
@@ -17,7 +17,7 @@ class PerfectMetamodel:
     def predict(self, X: pd.DataFrame) -> np.ndarray:
         if self.complete_y[X.index] is None:
             raise MalformedExperimentError("Dummy can't be used for new datapoints")
-        return self.complete_y[X.index].to_numpy()
+        return self.complete_y[X.index].to_numpy(dtype=np.float)
 
     def predict_proba(self, X: pd.DataFrame) -> np.ndarray:
         raise AssertionError("Perfect metamodel can't be used with probabilities. ")
